+++ src/main/java/org/truffleruby/cext/CExtNodes.java
@@ -21,10 +21,10 @@ import org.jcodings.Encoding;
 import org.jcodings.specific.UTF8Encoding;
 import org.truffleruby.Layouts;
 import org.truffleruby.RubyContext;
-import org.truffleruby.builtins.CoreModule;
 import org.truffleruby.builtins.CoreMethod;
 import org.truffleruby.builtins.CoreMethodArrayArgumentsNode;
 import org.truffleruby.builtins.CoreMethodNode;
+import org.truffleruby.builtins.CoreModule;
 import org.truffleruby.builtins.Primitive;
 import org.truffleruby.builtins.PrimitiveArrayArgumentsNode;
 import org.truffleruby.builtins.YieldingCoreMethodNode;
@@ -33,6 +33,7 @@ import org.truffleruby.cext.CExtNodesFactory.StringToNativeNodeGen;
 import org.truffleruby.core.CoreLibrary;
 import org.truffleruby.core.MarkingService.ExtensionCallStack;
 import org.truffleruby.core.MarkingServiceNodes;
+import org.truffleruby.core.RaiseIfFrozenNode;
 import org.truffleruby.core.array.ArrayHelpers;
 import org.truffleruby.core.array.ArrayOperationNodes;
 import org.truffleruby.core.array.ArrayStrategy;
@@ -77,7 +78,6 @@ import org.truffleruby.language.methods.InternalMethod;
 import org.truffleruby.language.objects.AllocateObjectNode;
 import org.truffleruby.language.objects.InitializeClassNode;
 import org.truffleruby.language.objects.InitializeClassNodeGen;
-import org.truffleruby.language.objects.IsFrozenNode;
 import org.truffleruby.language.objects.MetaClassNode;
 import org.truffleruby.language.objects.ReadObjectFieldNode;
 import org.truffleruby.language.objects.WriteObjectFieldNode;
@@ -639,11 +639,10 @@ public class CExtNodes {
     @CoreMethod(names = "rb_check_frozen", onSingleton = true, required = 1)
     public abstract static class CheckFrozenNode extends CoreMethodArrayArgumentsNode {
 
-        @Child private IsFrozenNode isFrozenNode = IsFrozenNode.create();
-
         @Specialization
-        protected boolean rb_check_frozen(Object object) {
-            isFrozenNode.raiseIfFrozen(object);
+        protected boolean rb_check_frozen(Object object,
+                @Cached RaiseIfFrozenNode raiseIfFrozenNode) {
+            raiseIfFrozenNode.execute(object);
             return true;
         }