+++ src/main/java/org/truffleruby/core/regexp/MatchDataNodes.java
@@ -48,6 +48,7 @@ import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.dsl.ImportStatic;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.object.DynamicObject;
 import com.oracle.truffle.api.profiles.BranchProfile;
@@ -55,6 +56,8 @@ import com.oracle.truffle.api.profiles.ConditionProfile;
 
 @CoreModule(value = "MatchData", isClass = true)
 public abstract class MatchDataNodes {
+
+    static final String UNINITIALIZED_MESSAGE = "uninitialized Match";
 
     @TruffleBoundary
     public static Object begin(RubyContext context, DynamicObject matchData, int index) {
@@ -201,6 +204,7 @@ public abstract class MatchDataNodes {
             lowerFixnum = { 1, 2 },
             taintFrom = 0,
             argumentNames = { "index_start_range_or_name", "length" })
+    @ImportStatic(MatchDataGuards.class)
     public abstract static class GetIndexNode extends CoreMethodArrayArgumentsNode {
 
         @Child private ToIntNode toIntNode;
@@ -215,7 +219,7 @@ public abstract class MatchDataNodes {
 
         public abstract Object executeGetIndex(Object matchData, Object index, Object length);
 
-        @Specialization
+        @Specialization(guards = "isInitialized(matchData)")
         protected Object getIndex(DynamicObject matchData, int index, NotProvided length,
                 @Cached("createBinaryProfile()") ConditionProfile normalizedIndexProfile,
                 @Cached("createBinaryProfile()") ConditionProfile indexOutOfBoundsProfile,
@@ -242,7 +246,7 @@ public abstract class MatchDataNodes {
             }
         }
 
-        @Specialization
+        @Specialization(guards = "isInitialized(matchData)")
         protected Object getIndex(DynamicObject matchData, int index, int length) {
             // TODO BJF 15-May-2015 Need to handle negative indexes and lengths and out of bounds
             final Object[] values = getValuesNode.execute(matchData);
@@ -253,6 +257,7 @@ public abstract class MatchDataNodes {
 
         @Specialization(
                 guards = {
+                        "isInitialized(matchData)",
                         "isRubySymbol(cachedIndex)",
                         "name != null",
                         "getRegexp(matchData) == regexp",
@@ -272,18 +277,23 @@ public abstract class MatchDataNodes {
             }
         }
 
-        @Specialization(guards = "isRubySymbol(index)")
+        @Specialization(guards = { "isInitialized(matchData)", "isRubySymbol(index)" })
         protected Object getIndexSymbol(DynamicObject matchData, DynamicObject index, NotProvided length,
                 @Cached BranchProfile errorProfile) {
             return executeGetIndex(matchData, getBackRefFromSymbol(matchData, index), NotProvided.INSTANCE);
         }
 
-        @Specialization(guards = "isRubyString(index)")
+        @Specialization(guards = { "isInitialized(matchData)", "isRubyString(index)" })
         protected Object getIndexString(DynamicObject matchData, DynamicObject index, NotProvided length) {
             return executeGetIndex(matchData, getBackRefFromString(matchData, index), NotProvided.INSTANCE);
         }
 
-        @Specialization(guards = { "!isRubySymbol(index)", "!isRubyString(index)", "!isIntRange(index)" })
+        @Specialization(
+                guards = {
+                        "isInitialized(matchData)",
+                        "!isRubySymbol(index)",
+                        "!isRubyString(index)",
+                        "!isIntRange(index)" })
         protected Object getIndex(DynamicObject matchData, Object index, NotProvided length) {
             if (toIntNode == null) {
                 CompilerDirectives.transferToInterpreterAndInvalidate();
@@ -294,7 +304,7 @@ public abstract class MatchDataNodes {
         }
 
         @TruffleBoundary
-        @Specialization(guards = "isIntRange(range)")
+        @Specialization(guards = { "isInitialized(matchData)", "isIntRange(range)" })
         protected Object getIndex(DynamicObject matchData, DynamicObject range, NotProvided len) {
             final Object[] values = getValuesNode.execute(matchData);
             final int normalizedIndex = ArrayOperations
@@ -306,6 +316,13 @@ public abstract class MatchDataNodes {
 
             final Object[] store = Arrays.copyOfRange(values, normalizedIndex, normalizedIndex + length);
             return createArray(store, length);
+        }
+
+        @Specialization(guards = "!isInitialized(matchData)")
+        protected DynamicObject uninitialized(DynamicObject matchData, Object index, NotProvided length) {
+            throw new RaiseException(
+                    getContext(),
+                    coreExceptions().typeError(MatchDataNodes.UNINITIALIZED_MESSAGE, this));
         }
 
         @TruffleBoundary
@@ -454,14 +471,22 @@ public abstract class MatchDataNodes {
     }
 
     @CoreMethod(names = "captures")
+    @ImportStatic(MatchDataGuards.class)
     public abstract static class CapturesNode extends CoreMethodArrayArgumentsNode {
 
         @Child private ValuesNode valuesNode = ValuesNode.create();
 
-        @Specialization
+        @Specialization(guards = "isInitialized(matchData)")
         protected DynamicObject toA(VirtualFrame frame, DynamicObject matchData) {
             Object[] objects = getCaptures(valuesNode.execute(matchData));
             return createArray(objects, objects.length);
+        }
+
+        @Specialization(guards = "!isInitialized(matchData)")
+        protected DynamicObject uninitialized(DynamicObject matchData) {
+            throw new RaiseException(
+                    getContext(),
+                    coreExceptions().typeError(MatchDataNodes.UNINITIALIZED_MESSAGE, this));
         }
 
         private static Object[] getCaptures(Object[] values) {
@@ -492,9 +517,10 @@ public abstract class MatchDataNodes {
 
     @NonStandard
     @CoreMethod(names = "byte_begin", required = 1, lowerFixnum = 1)
+    @ImportStatic(MatchDataGuards.class)
     public abstract static class ByteBeginNode extends CoreMethodArrayArgumentsNode {
 
-        @Specialization(guards = "inBounds(matchData, index)")
+        @Specialization(guards = { "isInitialized(matchData)", "inBounds(matchData, index)" })
         protected Object byteBegin(DynamicObject matchData, int index) {
             int b = Layouts.MATCH_DATA.getRegion(matchData).beg[index];
             if (b < 0) {
@@ -504,6 +530,13 @@ public abstract class MatchDataNodes {
             }
         }
 
+        @Specialization(guards = "!isInitialized(matchData)")
+        protected DynamicObject uninitialized(DynamicObject matchData, int index) {
+            throw new RaiseException(
+                    getContext(),
+                    coreExceptions().typeError(MatchDataNodes.UNINITIALIZED_MESSAGE, this));
+        }
+
         protected boolean inBounds(DynamicObject matchData, int index) {
             return index >= 0 && index < Layouts.MATCH_DATA.getRegion(matchData).numRegs;
         }
@@ -511,9 +544,10 @@ public abstract class MatchDataNodes {
 
     @NonStandard
     @CoreMethod(names = "byte_end", required = 1, lowerFixnum = 1)
+    @ImportStatic(MatchDataGuards.class)
     public abstract static class ByteEndNode extends CoreMethodArrayArgumentsNode {
 
-        @Specialization(guards = "inBounds(matchData, index)")
+        @Specialization(guards = { "isInitialized(matchData)", "inBounds(matchData, index)" })
         protected Object byteEnd(DynamicObject matchData, int index) {
             int e = Layouts.MATCH_DATA.getRegion(matchData).end[index];
             if (e < 0) {
@@ -523,24 +557,39 @@ public abstract class MatchDataNodes {
             }
         }
 
+        @Specialization(guards = "!isInitialized(matchData)")
+        protected DynamicObject uninitialized(DynamicObject matchData, int index) {
+            throw new RaiseException(
+                    getContext(),
+                    coreExceptions().typeError(MatchDataNodes.UNINITIALIZED_MESSAGE, this));
+        }
+
         protected boolean inBounds(DynamicObject matchData, int index) {
             return index >= 0 && index < Layouts.MATCH_DATA.getRegion(matchData).numRegs;
         }
     }
 
     @CoreMethod(names = { "length", "size" })
+    @ImportStatic(MatchDataGuards.class)
     public abstract static class LengthNode extends CoreMethodArrayArgumentsNode {
 
         @Child private ValuesNode getValues = ValuesNode.create();
 
-        @Specialization
+        @Specialization(guards = "isInitialized(matchData)")
         protected int length(DynamicObject matchData) {
             return getValues.execute(matchData).length;
         }
 
+        @Specialization(guards = "!isInitialized(matchData)")
+        protected DynamicObject uninitialized(DynamicObject matchData) {
+            throw new RaiseException(
+                    getContext(),
+                    coreExceptions().typeError(MatchDataNodes.UNINITIALIZED_MESSAGE, this));
+        }
     }
 
     @CoreMethod(names = "pre_match", taintFrom = 0)
+    @ImportStatic(MatchDataGuards.class)
     public abstract static class PreMatchNode extends CoreMethodArrayArgumentsNode {
 
         @Child private RopeNodes.SubstringNode substringNode = RopeNodes.SubstringNode.create();
@@ -548,7 +597,7 @@ public abstract class MatchDataNodes {
 
         public abstract DynamicObject execute(DynamicObject matchData);
 
-        @Specialization
+        @Specialization(guards = "isInitialized(matchData)")
         protected Object preMatch(DynamicObject matchData) {
             DynamicObject source = Layouts.MATCH_DATA.getSource(matchData);
             Rope sourceRope = StringOperations.rope(source);
@@ -560,9 +609,17 @@ public abstract class MatchDataNodes {
                     .allocate(Layouts.BASIC_OBJECT.getLogicalClass(source), Layouts.STRING.build(false, false, rope));
             return string;
         }
+
+        @Specialization(guards = "!isInitialized(matchData)")
+        protected DynamicObject uninitialized(DynamicObject matchData) {
+            throw new RaiseException(
+                    getContext(),
+                    coreExceptions().typeError(MatchDataNodes.UNINITIALIZED_MESSAGE, this));
+        }
     }
 
     @CoreMethod(names = "post_match", taintFrom = 0)
+    @ImportStatic(MatchDataGuards.class)
     public abstract static class PostMatchNode extends CoreMethodArrayArgumentsNode {
 
         @Child private RopeNodes.SubstringNode substringNode = RopeNodes.SubstringNode.create();
@@ -570,7 +627,7 @@ public abstract class MatchDataNodes {
 
         public abstract DynamicObject execute(DynamicObject matchData);
 
-        @Specialization
+        @Specialization(guards = "isInitialized(matchData)")
         protected Object postMatch(DynamicObject matchData) {
             DynamicObject source = Layouts.MATCH_DATA.getSource(matchData);
             Rope sourceRope = StringOperations.rope(source);
@@ -582,21 +639,37 @@ public abstract class MatchDataNodes {
                     .allocate(Layouts.BASIC_OBJECT.getLogicalClass(source), Layouts.STRING.build(false, false, rope));
             return string;
         }
+
+        @Specialization(guards = "!isInitialized(matchData)")
+        protected DynamicObject uninitialized(DynamicObject matchData) {
+            throw new RaiseException(
+                    getContext(),
+                    coreExceptions().typeError(MatchDataNodes.UNINITIALIZED_MESSAGE, this));
+        }
     }
 
     @CoreMethod(names = "to_a")
+    @ImportStatic(MatchDataGuards.class)
     public abstract static class ToANode extends CoreMethodArrayArgumentsNode {
 
         @Child ValuesNode valuesNode = ValuesNode.create();
 
-        @Specialization
+        @Specialization(guards = "isInitialized(matchData)")
         protected DynamicObject toA(DynamicObject matchData) {
             Object[] objects = ArrayUtils.copy(valuesNode.execute(matchData));
             return createArray(objects, objects.length);
         }
+
+        @Specialization(guards = "!isInitialized(matchData)")
+        protected DynamicObject uninitialized(DynamicObject matchData) {
+            throw new RaiseException(
+                    getContext(),
+                    coreExceptions().typeError(MatchDataNodes.UNINITIALIZED_MESSAGE, this));
+        }
     }
 
     @CoreMethod(names = "regexp")
+    @ImportStatic(MatchDataGuards.class)
     public abstract static class RegexpNode extends CoreMethodArrayArgumentsNode {
 
         public static RegexpNode create() {
@@ -605,7 +678,7 @@ public abstract class MatchDataNodes {
 
         public abstract DynamicObject executeGetRegexp(DynamicObject matchData);
 
-        @Specialization
+        @Specialization(guards = "isInitialized(matchData)")
         protected DynamicObject regexp(DynamicObject matchData,
                 @Cached("createBinaryProfile()") ConditionProfile profile,
                 @Cached("createPrivate()") CallDispatchHeadNode stringToRegexp) {
@@ -623,18 +696,49 @@ public abstract class MatchDataNodes {
             }
         }
 
+        @Specialization(guards = "!isInitialized(matchData)")
+        protected DynamicObject uninitialized(DynamicObject matchData) {
+            throw new RaiseException(
+                    getContext(),
+                    coreExceptions().typeError(MatchDataNodes.UNINITIALIZED_MESSAGE, this));
+        }
+
     }
 
     @CoreMethod(names = "__allocate__", constructor = true, visibility = Visibility.PRIVATE)
     public abstract static class AllocateNode extends UnaryCoreMethodNode {
 
-        // MatchData can be allocated in MRI but it does not seem to be any useful
         @TruffleBoundary
         @Specialization
-        protected DynamicObject allocate(DynamicObject rubyClass) {
-            throw new RaiseException(getContext(), coreExceptions().typeErrorAllocatorUndefinedFor(rubyClass, this));
+        protected DynamicObject allocate(DynamicObject rubyClass,
+                @Cached AllocateObjectNode allocateNode) {
+            return allocateNode.allocate(rubyClass, Layouts.MATCH_DATA.build(null, null, null, null));
         }
 
+    }
+
+    @CoreMethod(names = "initialize_copy", required = 1, raiseIfFrozenSelf = true)
+    public abstract static class InitializeCopyNode extends CoreMethodArrayArgumentsNode {
+
+        @Specialization
+        protected DynamicObject initializeCopy(DynamicObject self, DynamicObject from) {
+            if (self == from) {
+                return self;
+            }
+
+            if (Layouts.BASIC_OBJECT.getLogicalClass(self) != Layouts.BASIC_OBJECT.getLogicalClass(from)) {
+                throw new RaiseException(
+                        getContext(),
+                        coreExceptions().typeError("initialize_copy should take same class object", this));
+            }
+
+
+            Layouts.MATCH_DATA.setSource(self, Layouts.MATCH_DATA.getSource(from));
+            Layouts.MATCH_DATA.setRegexp(self, Layouts.MATCH_DATA.getRegexp(from));
+            Layouts.MATCH_DATA.setRegion(self, Layouts.MATCH_DATA.getRegion(from));
+            Layouts.MATCH_DATA.setCharOffsets(self, Layouts.MATCH_DATA.getCharOffsets(from));
+            return self;
+        }
     }
 
     @Primitive(name = "match_data_get_source")
@@ -643,6 +747,15 @@ public abstract class MatchDataNodes {
         @Specialization
         protected DynamicObject getSource(DynamicObject matchData) {
             return Layouts.MATCH_DATA.getSource(matchData);
+        }
+    }
+
+    @Primitive(name = "match_data_initialized?")
+    public abstract static class InitializedNode extends PrimitiveArrayArgumentsNode {
+
+        @Specialization
+        protected boolean isInitialized(DynamicObject matchData) {
+            return MatchDataGuards.isInitialized(matchData);
         }
     }